requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-countries==7.5.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
idna==3.4
phonenumbers==8.13.6
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
python-decouple==3.7
pytz==2022.7.1
pyuca==1.2
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
